split regression tests
